tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.reference.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.host.reference</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.reference</h1>
<p class="desc">canvas refers back to its canvas</p>


<script>
var t = async_test("canvas refers back to its canvas");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertSame(ctx.canvas, canvas, "ctx.canvas", "canvas");
  t.done();

});
</script>
